backend-api-service/requirements.txt
fastapi==0.104.1
pydantic
pydantic[email]
email-validator
python-dotenv==1.0.1
requests==2.31.0
uvicorn==0.24.0
websockets==11.0.3
elevenlabs=1.54.0
certifi==2023.11.17
# SQLite-related packages
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.25.2
loguru==0.7.2
aiohttp==3.11.14
elevenlabs
# Optional - for public URL tunneling
pyngrok
